requirements.txt
streamlit>=1.28.0
stripe>=5.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
email-validator>=2.0.0
flask>=2.3.0